+++ R/MassIVE_functions.R
@@ -52,16 +52,16 @@
 #'   Parameter `fileName` allows to specify names of selected data files to
 #'   sync.
 #'
-#' - `massive_data_download()`: download files from the MassIVE repository for a
+#' - `massive_download_file()`: download files from the MassIVE repository for a
 #'   specified MassIVE dataset. Use `pattern` to filter files by name using a
 #'   regular expression (downloads all files by default). Use `fileName` to
 #'   specify one or more exact file names to download. Use `path` to set the
 #'   destination directory for downloaded files.
 #'
-#' - `massive_param_file_parse()`: download and parse the `params.xml` files of
-#'   the data set. The function return a data.frame or a list of data.frame with
-#'   2 columns (Parameter Name, Value). Use `fileName` to parse additional `xml`
-#'   files in the data.set.
+#' - `massive_param_file()`: download and parse the `params.xml` files of
+#'   the data set. The function return a `data.frame` or a `list` of
+#'   `data.frame` with 2 columns (Parameter Name, Value). Use `fileName` to
+#'   parse additional `xml` files in the data.set.
 #'
 #' - `massive_delete_cache()`: removes all local content for the MassIVE
 #'   data set with ID `massiveId`. This will delete eventually present
@@ -110,10 +110,10 @@
 #'     defining the names of specific data files of a data set that should be
 #'     downloaded and cached.
 #'
-#' @param path for `massive_data_download()`: optional `character` defining the
+#' @param path for `massive_download_file()`: optional `character` defining the
 #'     directory where download the files.
 #'
-#' @param overwrite for `massive_data_download()`: `logical(1)` whether
+#' @param overwrite for `massive_download_file()`: `logical(1)` whether
 #'     existing files should be overwritten. Defaults to `FALSE`, in which
 #'     case files that already exist in `path` are skipped.
 #'
@@ -142,7 +142,7 @@
 #' mzMLfiles
 #'
 #' ## Download parameter file for the data set MSV000080547
-#' massive_data_download("MSV000080547", pattern = "params.xml",
+#' massive_download_file("MSV000080547", pattern = "params.xml",
 #'                       path = tempdir())
 #'
 NULL
@@ -222,7 +222,7 @@ massive_list_files <- function(x = character(), pattern = NULL) {
 #' @importFrom utils capture.output URLencode download.file
 #'
 #' @export
-massive_data_download <- function(massiveId = character(), pattern = "*",
+massive_download_file <- function(massiveId = character(), pattern = "*",
                                   fileName = character(), path = "./",
                                   overwrite = FALSE){
     if (!length(massiveId))
@@ -245,20 +245,19 @@ massive_data_download <- function(massiveId = character(), pattern = "*",
     ## Create dir if not exist
     if (!dir.exists(path)) {
         dir.create(path, recursive = TRUE)
-        message("Created directory: ", path)
     }
 
     ## Update the Volume if files are in ccms_peak folder
     ## ccms_peak is in volume z01 for all the project
     api_z_volume = "ftp://massive-ftp.ucsd.edu/z01/"
-    ffiles <- sapply(dfiles,
+    ffiles <- vapply(dfiles,
                      function(f) {
                          u <- ifelse(grepl("^ccms_peak", f),
                                      paste0(api_z_volume, massiveId, "/", f),
                                      paste0(fpath, f))
                          ## URLencode for file name with spaces
                          URLencode(u)
-                     }, USE.NAMES = FALSE)
+                     }, FUN.VALUE = character(1), USE.NAMES = FALSE)
 
     ## Save files in the folder
     pb <- progress_bar$new(format = paste0("[:bar] :current/:",
@@ -286,7 +285,7 @@ massive_data_download <- function(massiveId = character(), pattern = "*",
 #' @importFrom xml2 read_xml xml_find_all xml_attrs xml_text
 #'
 #' @export
-massive_param_file_parse <- function(massiveId = character(),
+massive_param_file <- function(massiveId = character(),
                                      fileName = "params.xml") {
     if (!length(massiveId))
         stop("No MassIVE data set ID provided with parameter 'massiveId'")